Fix undeterministic list item duplication
Sometimes lists were being parsed with more instances of an element than
expected.

This behavior wasn't only unwanted but also undeterministic: it only
happened at times. There is some indeterminism expected from the YAML
parser (since the YAML spec doesn't make keys sensible to sorting), and
in this case it was the source of the problem.